fix: hide tabBar
