+++ App.js
@@ -1,43 +1,22 @@
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { HomeScreen } from './src/screens/home/HomeScreen';
-import { ProfileScreen } from './src/screens/profile/ProfileScreen';
-import { Ionicons } from '@expo/vector-icons';
-import { COLORS, SPACING } from './src/utils/theme'
 import { MainStackScreen } from './src/screens/event-list/MainStackScreen';
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { EventDetailScreen } from "./src/screens/event-detail/EventDetailScreen";
 
-const Tab = createBottomTabNavigator();
-
-const TAB_ICON = {
-  Home: 'newspaper',
-  Eventos: 'calendar',
-  Perfil: 'happy'
-}
-
-const screenOptions = ({ route }) => {
-  const iconName = TAB_ICON[route.name]
-  return {
-    tabBarIcon: ({ size, color }) => (
-      <Ionicons name={iconName} size={size} color={color} />
-    ),
-    tabBarActiveTintColor: COLORS.success,
-    tabBarInactiveTintColor: COLORS.inactive,
-    headerShown: false,
-    tabBarStyle: styles.tabBar
-  }
-}
+const EventListStack = createNativeStackNavigator()
 
 export default function App() {
   return (
     <>
       <NavigationContainer>
-        <Tab.Navigator screenOptions={screenOptions}>
-          <Tab.Screen name="Home" component={HomeScreen} />
-          <Tab.Screen name="Eventos" component={MainStackScreen} />
-          <Tab.Screen name="Perfil" component={ProfileScreen} />
-        </Tab.Navigator>
+        <EventListStack.Navigator screenOptions={{ headerShown: false }}>
+
+          <EventListStack.Screen name="Main" component={MainStackScreen} />
+
+          <EventListStack.Screen name="EventDetail" component={EventDetailScreen}></EventListStack.Screen>
+
+        </EventListStack.Navigator>
       </NavigationContainer>
       <StatusBar style="auto" />
     </>
@@ -45,11 +24,4 @@ export default function App() {
   );
 }
 
-const styles = StyleSheet.create({
-  tabBar:{
-    height: SPACING.xxxl,
-    paddingBottom: SPACING.xs,
-    paddingTop: SPACING.xs,
-    backgroundColor: COLORS.black
-  }
-})
+
